requirements.txt
PyOpenGL==3.1.7
glfw==2.7.0
